requirements.txt
pymongo>=4.6.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.3.0
rich>=13.7.0
prompt_toolkit>=3.0.43
tqdm>=4.66.0
requests>=2.31.0